quartz instrumentation is divided into modules
