+++ apm-agent-plugins/apm-quartz-job-plugin/quartz-1/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,0 +1,1 @@
+co.elastic.apm.agent.quartzjob.Quartz2JobTransactionNameInstrumentationTest
