+++ apm-agent-plugins/apm-quartz-job-plugin/quartz-1/src/test/java/co/elastic/apm/agent/quartzjob/Quartz1JobTransactionNameInstrumentationTest.java
@@ -1,0 +1,4 @@
+package co.elastic.apm.agent.quartzjob;
+
+public class Quartz1JobTransactionNameInstrumentationTest extends AbstractJobTransactionNameInstrumentationTest {
+}
