+++ apm-agent-plugins/apm-quartz-job-plugin/quartz-2/src/test/java/co/elastic/apm/agent/quartzjob/Quartz2JobTransactionNameInstrumentationTest.java
@@ -1,0 +1,4 @@
+package co.elastic.apm.agent.quartzjob;
+
+public class Quartz2JobTransactionNameInstrumentationTest extends AbstractJobTransactionNameInstrumentationTest {
+}
