+++ apm-agent-plugins/apm-quartz-job-plugin/quartz-common/src/main/java/co/elastic/apm/agent/quartzjob/JobExecutionContextHandler.java
@@ -1,0 +1,13 @@
+package co.elastic.apm.agent.quartzjob;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+public interface JobExecutionContextHandler<T> {
+
+    @Nullable
+    String getJobDetailKey(@Nonnull T jobExecutionContext);
+
+    @Nullable
+    Object getResult(@Nonnull T jobExecutionContext);
+}
