--- apm-agent-plugins/apm-quartz-job-plugin/src/main/java/co/elastic/apm/agent/quartzjob/JobTransactionNameInstrumentation.java
@@ -1,75 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package co.elastic.apm.agent.quartzjob;
-
-import co.elastic.apm.agent.bci.TracerAwareInstrumentation;
-import co.elastic.apm.agent.impl.ElasticApmTracer;
-import co.elastic.apm.agent.impl.stacktrace.StacktraceConfiguration;
-import net.bytebuddy.description.NamedElement;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-import net.bytebuddy.matcher.ElementMatchers;
-
-import java.util.Arrays;
-import java.util.Collection;
-
-import static co.elastic.apm.agent.bci.bytebuddy.CustomElementMatchers.isInAnyPackage;
-import static net.bytebuddy.matcher.ElementMatchers.declaresMethod;
-import static net.bytebuddy.matcher.ElementMatchers.hasSuperType;
-import static net.bytebuddy.matcher.ElementMatchers.isInterface;
-import static net.bytebuddy.matcher.ElementMatchers.nameStartsWith;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-
-public class JobTransactionNameInstrumentation extends TracerAwareInstrumentation {
-    public static final String TRANSACTION_TYPE = "scheduled";
-    public static final String INSTRUMENTATION_TYPE = "quartz";
-
-    private final Collection<String> applicationPackages;
-
-    public JobTransactionNameInstrumentation(ElasticApmTracer tracer) {
-        applicationPackages = tracer.getConfig(StacktraceConfiguration.class).getApplicationPackages();
-    }
-
-    @Override
-    public ElementMatcher<? super TypeDescription> getTypeMatcher() {
-        return isInAnyPackage(applicationPackages, ElementMatchers.<NamedElement>none())
-            .or(nameStartsWith("org.quartz.job"))
-            .and(hasSuperType(named("org.quartz.Job")))
-            .and(declaresMethod(getMethodMatcher()));
-    }
-
-    @Override
-    public ElementMatcher<? super MethodDescription> getMethodMatcher() {
-        return named("execute").or(named("executeInternal"))
-            .and(takesArgument(0, named("org.quartz.JobExecutionContext").and(isInterface())));
-    }
-
-    @Override
-    public Collection<String> getInstrumentationGroupNames() {
-        return Arrays.asList(INSTRUMENTATION_TYPE);
-    }
-
-    @Override
-    public String getAdviceClassName() {
-        return "co.elastic.apm.agent.quartzjob.JobTransactionNameAdvice";
-    }
-
-}
