--- apm-agent-plugins/apm-quartz-job-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,1 +1,0 @@
-co.elastic.apm.agent.quartzjob.JobTransactionNameInstrumentation
